tiktok/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
celery[redis]>=5.3.0
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
httpx>=0.27.0
openai>=1.12.0
elevenlabs>=1.0.0
mutagen>=1.47.0
apscheduler>=3.10.0
loguru>=0.7.0
pillow>=10.2.0
ffmpeg-python>=0.2.0
websockets>=12.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
